chat-assistant/backend/llm-service/requirements.txt
fastapi
grpcio
grpcio-tools
uvicorn
openai
asyncio
